alan/requirements.txt
albumentations==1.2.1
evaluate==0.4.0
numpy==1.22.4
opencv_python==4.7.0.72
pandas==1.4.4
Pillow==9.4.0
rasterio==1.3.6
scikit_learn==1.2.2
torch==1.13.1+cu116
tqdm==4.65.0
transformers==4.27.3
GDAL==3.3.2
matplotlib==3.7.1
osgeo==0.0.1
scikit_image==0.19.3
scipy==1.10.1
skimage==0.0
tensorflow==2.11.0

